Changed type from long to int - should be enough

+++ src/LemonTree.Pipeline.Tools/Database/JetDatabase.cs
@@ -66,16 +66,16 @@ namespace LemonTree.Pipeline.Tools.Database
             return RecordCount;
         }
 
-        public long RunSQLQueryScalar(string sql)
+        public int RunSQLQueryScalar(string sql)
         {
-            long RecordCount = 0;
+            int RecordCount = 0;
 
             using (var cn = new OleDbConnection { ConnectionString = _builder.ConnectionString })
             {
                 using (var cmd = new OleDbCommand { CommandText = sql, Connection = cn })
                 {
                     cn.Open();
-                    RecordCount = Convert.ToInt64(cmd.ExecuteScalar());
+                    RecordCount = Convert.ToInt32(cmd.ExecuteScalar());
 
                 }
             }
