+++ src/LemonTree.Pipeline.Tools/Database/SqLiteDatabase.cs
@@ -78,16 +78,16 @@ namespace LemonTree.Pipeline.Tools.Database
             return RecordCount;
         }
 
-        public long RunSQLQueryScalar(string sql)
+        public int RunSQLQueryScalar(string sql)
         {
-            long RecordCount = 0;
+            int RecordCount = 0;
 
             using (var cn = new SQLiteConnection { ConnectionString = _builder.ConnectionString })
             {
                 using (var cmd = new SQLiteCommand { CommandText = sql, Connection = cn })
                 {
                     cn.Open();
-                    RecordCount = (long)cmd.ExecuteScalar();
+                    RecordCount = (int)cmd.ExecuteScalar();
 
                 }
             }
